requirements.txt
web3==6.0.0
python-dotenv==1.0.0
aiohttp==3.8.1
Flask==2.2.2
Flask-SocketIO==5.3.2
eventlet==0.33.0
